+++ R/AeRobiology.R
@@ -8,4 +8,4 @@
 NULL
 
 ## quiets concerns of R CMD check re: the .'s that appear in pipelines
-if(getRversion() >= "2.15.1")  utils::globalVariables(c(".", ":=", "DOY", "Max", "Mean", "Min", "Risk", "Year", "en.jd", "j.days", "jd", "opacity", "phen", "pk.jd", "pollen", "seasons", "size", "sm.ps", "st.jd", "strokeWidth", "value", "week", "wt", "x", "variable", "Hour", "location", "percent"),add=TRUE)
+if(getRversion() >= "2.15.1")  utils::globalVariables(c(".", ":=", "DOY", "Max", "Mean", "Min", "Risk", "Year", "en.jd", "j.days", "jd", "opacity", "phen", "pk.jd", "pollen", "seasons", "size", "sm.ps", "st.jd", "strokeWidth", "value", "week", "wt", "x", "variable", "Hour", "location", "percent", "month.label", "sum.pollen"),add=TRUE)